requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
Django==2.2.2
django-redis-cache==2.0.0
djangorestframework==3.9.4
pytz==2019.1
redis==3.2.1
sqlparse==0.3.0
toml==0.10.0
